requirements.txt
asgiref==3.5.0
Django==4.0.2
django-ckeditor==6.2.0
django-js-asset==2.0.0
fontawesomefree==6.1.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.2.0
